requirements.txt
Django>=3.2
djangorestframework
django-rest-authtoken==2.1.4
pymongo
dnspython
django-extensions
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
python-dotenv
whitenoise==6.7.0